1)allergy is linked

+++ lib/controllers/backend_controller.dart
@@ -1,9 +1,7 @@
 import 'dart:convert';
 import 'dart:convert' as convert;
-
-//import 'dart:html';
-import 'package:flutter/services.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+import 'package:foodcam_frontend/models/allergy.dart';
 import 'package:foodcam_frontend/models/category.dart';
 import 'package:foodcam_frontend/models/collection.dart';
 import 'package:foodcam_frontend/models/ingredient.dart';
@@ -53,10 +51,6 @@ class BackEndController {
     return recipes;
   }
 
-  // Future<Recipe> generateRandomRecipe(String categoryName,String langcode){
-
-  // }
-
   Future<List<Collection>> getUserCollections(String langCode) async {
     final SharedPreferences _sharedPreferences =
         await SharedPreferences.getInstance();
@@ -77,9 +71,7 @@ class BackEndController {
           'Authorization': 'Bearer $accessToken',
         },
       );
-      //TODO: check response and send refresh token if the access token is exired
       var _responseJson;
-      //print(response.body);
       if (response.body.contains('Given token not valid')) {
         //print('here');
         final String newAccessToken = await this.refreshToken(refreshToken!);
@@ -90,7 +82,6 @@ class BackEndController {
             'Authorization': 'Bearer $newAccessToken',
           },
         );
-        //print(newResponse);
         _responseJson = jsonDecode(
           utf8.decode(newResponse.bodyBytes),
         );
@@ -99,7 +90,6 @@ class BackEndController {
           utf8.decode(response.bodyBytes),
         );
       }
-      //print(_responseJson);
       final List<Collection> collections = [];
 
       for (final item in _responseJson) {
@@ -108,7 +98,6 @@ class BackEndController {
 
       return collections;
     } else {
-      //print("username is null");
       return [];
     }
   }
@@ -230,7 +219,6 @@ class BackEndController {
         await flutterSecureStorage.read(key: 'refresh_token');
     final SharedPreferences _sharedPreferences =
         await SharedPreferences.getInstance();
-    //print("Saeed");
     final String? userName = _sharedPreferences.getString('userName');
     if (userName != null) {
       final url = Uri.parse(
@@ -242,21 +230,32 @@ class BackEndController {
           'Authorization': 'Bearer $accessToken',
         },
       );
+      var _responseJson;
+      if (response.body.contains('Given token not valid')) {
+        final String newAccessToken = await this.refreshToken(refreshToken!);
+        final http.Response newResponse = await http.get(
+          url,
+          headers: {
+            'Content-Type': 'application/json; charset=UTF-8',
+            'Authorization': 'Bearer $newAccessToken',
+          },
+        );
+        _responseJson = jsonDecode(
+          utf8.decode(newResponse.bodyBytes),
+        );
+      } else {
+        _responseJson = jsonDecode(
+          utf8.decode(response.bodyBytes),
+        );
+      }
 
-      final _responseJson = jsonDecode(utf8.decode(response.bodyBytes));
-
-      //print(_responseJson);
       final List<Ingredient> ingredients = [];
 
       for (final item in _responseJson) {
-        //print(item);
         ingredients.add(Ingredient.fromMapBasket(item));
       }
-      // print("sae");
-      //print(ingredients);
       return ingredients;
     } else {
-      print("username is null");
       return [];
     }
   }
@@ -274,13 +273,23 @@ class BackEndController {
     if (userName != null) {
       final url = Uri.parse(
           "http://192.168.1.5:8000/UserIngredientView/?username=$userName&ingredient=$ingredientID");
-      await http.delete(
+      final http.Response response = await http.delete(
         url,
         headers: {
           'Content-Type': 'application/json; charset=UTF-8',
           'Authorization': 'Bearer $accessToken',
         },
       );
+      if (response.body.contains('Given token not valid')) {
+        final String newAccessToken = await this.refreshToken(refreshToken!);
+        await http.delete(
+          url,
+          headers: {
+            'Content-Type': 'application/json; charset=UTF-8',
+            'Authorization': 'Bearer $newAccessToken',
+          },
+        );
+      }
     }
   }
 
@@ -308,9 +317,25 @@ class BackEndController {
         'Authorization': 'Bearer $accessToken',
       },
     );
-    final _responseJson = jsonDecode(utf8.decode(response.bodyBytes));
-    print(_responseJson);
-    return response;
+    if (response.body.contains('Given token not valid')) {
+      print('expired');
+      final String newAccessToken = await this.refreshToken(refreshToken!);
+      final http.Response newResponse = await http.post(
+        url,
+        body: convert.jsonEncode(<String, dynamic>{
+          'username': userName,
+          'ingredient': ingredientId,
+        }),
+        headers: {
+          'Content-Type': 'application/json; charset=UTF-8',
+          'Authorization': 'Bearer $newAccessToken',
+        },
+      );
+      print(newResponse.body);
+      return newResponse;
+    } else {
+      return response;
+    }
   }
 
   Future<List<Map<String, dynamic>>> ingredientBasketSearch(
@@ -322,7 +347,6 @@ class BackEndController {
         await flutterSecureStorage.read(key: 'refresh_token');
     final SharedPreferences _sharedPreferences =
         await SharedPreferences.getInstance();
-    //print("Saeed");
     final String? userName = _sharedPreferences.getString('userName');
     if (userName != null) {
       final url = Uri.parse(
@@ -334,13 +358,21 @@ class BackEndController {
           'Authorization': 'Bearer $accessToken',
         },
       );
-
-      final _responseJson = jsonDecode(utf8.decode(response.bodyBytes));
-
-      print(_responseJson);
-      //print("Saeed");
+      var _responseJson;
+      if (response.body.contains('Given token not valid')) {
+        final String newAccessToken = await this.refreshToken(refreshToken!);
+        final http.Response newResponse = await http.get(
+          url,
+          headers: {
+            'Content-Type': 'application/json; charset=UTF-8',
+            'Authorization': 'Bearer $newAccessToken',
+          },
+        );
+        _responseJson = jsonDecode(utf8.decode(newResponse.bodyBytes));
+      } else {
+        _responseJson = jsonDecode(utf8.decode(response.bodyBytes));
+      }
       final List<Map<String, dynamic>> ingredients = [];
-
       for (final Map<String, dynamic> item in _responseJson) {
         final Ingredient ingredient = Ingredient.fromMapBasket(item);
         final bool isAdded = item['ingredient']['my_ingredient'];
@@ -349,11 +381,8 @@ class BackEndController {
           'isAdded': isAdded,
         });
       }
-      //print("Saeed");
-      //print(ingredients);
       return ingredients;
     } else {
-      print("username is null");
       return [];
     }
   }
@@ -382,7 +411,6 @@ class BackEndController {
         'Authorization': 'Bearer $accessToken',
       },
     );
-    //TODO: check response and send refresh token if the access token is exired
     if (response.body.contains('Given token not valid')) {
       final String newAccessToken = await this.refreshToken(refreshToken!);
       final http.Response newResponse = await http.post(
@@ -394,7 +422,7 @@ class BackEndController {
         }),
         headers: {
           'Content-Type': 'application/json; charset=UTF-8',
-          'Authorization': 'Bearer $accessToken',
+          'Authorization': 'Bearer $newAccessToken',
         },
       );
       return newResponse;
@@ -405,8 +433,6 @@ class BackEndController {
 
   Future<void> deleteCollection(String collectionName) async {
     final SharedPreferences _sharedPreferences =
-        await SharedPreferences.getInstance();
-    final SharedPreferences sharedPreferences =
         await SharedPreferences.getInstance();
     const FlutterSecureStorage flutterSecureStorage = FlutterSecureStorage();
     final String? accessToken =
@@ -456,7 +482,6 @@ class BackEndController {
         'Authorization': 'Bearer $accessToken',
       },
     );
-    //TODO: check response and send refresh token if the access token is exired
     var responseJson;
     if (response.body.contains('Given token not valid')) {
       final String newAccessToken = await this.refreshToken(refreshToken!);
@@ -546,6 +571,155 @@ class BackEndController {
           'Content-Type': 'application/json; charset=UTF-8',
           'Authorization': 'Bearer $newAccessToken',
         },
+      );
+    }
+  }
+
+  Future<List<Allergy>> getUserAllergies(String langCode) async {
+    final SharedPreferences sharedPreferences =
+        await SharedPreferences.getInstance();
+    const FlutterSecureStorage flutterSecureStorage = FlutterSecureStorage();
+    final String? accessToken =
+        await flutterSecureStorage.read(key: 'access_token');
+    final String? refreshToken =
+        await flutterSecureStorage.read(key: 'refresh_token');
+    final String userName = sharedPreferences.getString('userName')!;
+    final url = Uri.parse(
+        'http://192.168.1.5:8000/userPrefAllergy/?username=$userName&lang_code=$langCode');
+    final http.Response response = await http.get(url, headers: {
+      'Content-Type': 'application/json; charset=UTF-8',
+      'Authorization': 'Bearer $accessToken',
+    });
+    print(response.body);
+    var _responseJson;
+    if (response.body.contains('Given token not valid')) {
+      final String newAccessToken = await this.refreshToken(refreshToken!);
+      final http.Response newResponse = await http.get(url, headers: {
+        'Content-Type': 'application/json; charset=UTF-8',
+        'Authorization': 'Bearer $newAccessToken',
+      });
+      _responseJson = jsonDecode(
+        utf8.decode(newResponse.bodyBytes),
+      );
+    } else {
+      _responseJson = jsonDecode(
+        utf8.decode(response.bodyBytes),
+      );
+    }
+    final List<Allergy> allergies = [];
+    for (final item in _responseJson) {
+      allergies.add(Allergy.fromMap(item));
+    }
+    return allergies;
+  }
+
+  Future<List<Map<String, dynamic>>> getAllergies(String langCode) async {
+    final SharedPreferences sharedPreference =
+        await SharedPreferences.getInstance();
+    const FlutterSecureStorage flutterSecureStorage = FlutterSecureStorage();
+    final String? accessToken =
+        await flutterSecureStorage.read(key: 'access_token');
+    final String? refreshToken =
+        await flutterSecureStorage.read(key: 'refresh_token');
+    final String userName = sharedPreference.getString('userName')!;
+    final url = Uri.parse(
+        'http://192.168.1.5:8000/GetAllergies/?lang_code=$langCode&username=$userName');
+    final http.Response response = await http.get(url, headers: {
+      'Content-Type': 'application/json; charset=UTF-8',
+      'Authorization': 'Bearer $accessToken',
+    });
+    var _responseJson;
+    if (response.body.contains('Given token not valid')) {
+      final String newAccessToken = await this.refreshToken(refreshToken!);
+      final http.Response newResponse = await http.get(url, headers: {
+        'Content-Type': 'application/json; charset=UTF-8',
+        'Authorization': 'Bearer $newAccessToken',
+      });
+      _responseJson = jsonDecode(
+        utf8.decode(newResponse.bodyBytes),
+      );
+    } else {
+      _responseJson = jsonDecode(
+        utf8.decode(response.bodyBytes),
+      );
+    }
+    List<Map<String, dynamic>> allergies = [];
+    for (final item in _responseJson) {
+      final Allergy allergy = Allergy.fromMapAllAllergies(item);
+      final bool isAdded = item['isExist_In_my_Allergies'];
+      allergies.add({
+        'allergy': allergy,
+        'isAdded': isAdded,
+      });
+    }
+    return allergies;
+  }
+
+  Future<void> addAllergy(int allergyId) async {
+    final SharedPreferences sharedPreferences =
+        await SharedPreferences.getInstance();
+    const FlutterSecureStorage flutterSecureStorage = FlutterSecureStorage();
+    final String? accessToken =
+        await flutterSecureStorage.read(key: 'access_token');
+    final String? refreshToken =
+        await flutterSecureStorage.read(key: 'refresh_token');
+    final String userName = sharedPreferences.getString('userName')!;
+    final url = Uri.parse('http://192.168.1.5:8000/userPrefAllergy/');
+    final http.Response response = await http.post(
+      url,
+      headers: {
+        'Content-Type': 'application/json; charset=UTF-8',
+        'Authorization': 'Bearer $accessToken',
+      },
+      body: convert.jsonEncode(
+        <String, dynamic>{'username': userName, 'allergy': allergyId},
+      ),
+    );
+    if (response.body.contains('Given token not valid')) {
+      final String newAccessToken = await this.refreshToken(refreshToken!);
+      await http.post(
+        url,
+        headers: {
+          'Content-Type': 'application/json; charset=UTF-8',
+          'Authorization': 'Bearer $newAccessToken',
+        },
+        body: convert.jsonEncode(
+          <String, dynamic>{'username': userName, 'allergy': allergyId},
+        ),
+      );
+    }
+  }
+
+  Future<void> deleteAllergy(int allergyId) async {
+    final SharedPreferences sharedPreferences =
+        await SharedPreferences.getInstance();
+    const FlutterSecureStorage flutterSecureStorage = FlutterSecureStorage();
+    final String? accessToken =
+        await flutterSecureStorage.read(key: 'access_token');
+    final String? refreshToken =
+        await flutterSecureStorage.read(key: 'refresh_token');
+    final String userName = sharedPreferences.getString('userName')!;
+    final url = Uri.parse(
+        'http://192.168.1.5:8000/DeleteAllergy/?allergy=$allergyId&username=$userName');
+    final http.Response response = await http.delete(
+      url,
+      headers: {
+        'Content-Type': 'application/json; charset=UTF-8',
+        'Authorization': 'Bearer $accessToken',
+      },
+    );
+
+    if (response.body.contains('Given token not valid')) {
+      final String newAccessToken = await this.refreshToken(refreshToken!);
+      await http.post(
+        url,
+        headers: {
+          'Content-Type': 'application/json; charset=UTF-8',
+          'Authorization': 'Bearer $newAccessToken',
+        },
+        body: convert.jsonEncode(
+          <String, dynamic>{'username': userName, 'allergy': allergyId},
+        ),
       );
     }
   }
